7Modules/Melrick/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
aiosqlite==0.20.0
httpx==0.27.2
python-jose[cryptography]==3.3.0
python-socketio==5.11.4
passlib[bcrypt]==1.7.4
pydantic==2.12.0
greenlet==3.2.4
ortools>=9.10
